Finalize Restaurant page
